backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23

python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.0.1
Pillow==10.1.0
python-socketio==5.10.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1
python-dotenv==1.0.0
pymysql